Moved code to src folder
