+++ src/uepyscripts/tools/ue/generate_solution.py
